# ci: unflake "go perf"