--- a/internal/wasm/sdk/internal/wasm/pool.go
+++ b/internal/wasm/sdk/internal/wasm/pool.go
@@ -153,7 +153,7 @@ func (p *Pool) Release(vm *VM, metrics metrics.Metrics) {
 // are constructed in advance before touching the pool.  Returns
 // either ErrNotReady, ErrInvalidPolicy or ErrInternal if an error
 // occurs.
-func (p *Pool) SetPolicyData(policy []byte, data []byte) error {
+func (p *Pool) SetPolicyData(ctx context.Context, policy []byte, data []byte) error {
 	p.dataMtx.Lock()
 	defer p.dataMtx.Unlock()
 
@@ -196,7 +196,7 @@ func (p *Pool) SetPolicyData(policy []byte, data []byte) error {
 		return nil
 	}
 
-	err := p.setPolicyData(policy, data)
+	err := p.setPolicyData(ctx, policy, data)
 	if err != nil {
 		return fmt.Errorf("%v: %w", err, errors.ErrInternal)
 	}
@@ -207,31 +207,31 @@ func (p *Pool) SetPolicyData(policy []byte, data []byte) error {
 // SetDataPath will update the current data on the VMs by setting the value at the
 // specified path. If an error occurs the instance is still in a valid state, however
 // the data will not have been modified.
-func (p *Pool) SetDataPath(path []string, value interface{}) error {
+func (p *Pool) SetDataPath(ctx context.Context, path []string, value interface{}) error {
 	p.dataMtx.Lock()
 	defer p.dataMtx.Unlock()
 	return p.updateVMs(func(vm *VM, opts vmOpts) error {
-		return vm.SetDataPath(path, value)
+		return vm.SetDataPath(ctx, path, value)
 	})
 }
 
 // RemoveDataPath will update the current data on the VMs by removing the value at the
 // specified path. If an error occurs the instance is still in a valid state, however
 // the data will not have been modified.
-func (p *Pool) RemoveDataPath(path []string) error {
+func (p *Pool) RemoveDataPath(ctx context.Context, path []string) error {
 	p.dataMtx.Lock()
 	defer p.dataMtx.Unlock()
 	return p.updateVMs(func(vm *VM, _ vmOpts) error {
-		return vm.RemoveDataPath(path)
+		return vm.RemoveDataPath(ctx, path)
 	})
 }
 
 // setPolicyData reinitializes the VMs one at a time.
-func (p *Pool) setPolicyData(policy []byte, data []byte) error {
+func (p *Pool) setPolicyData(ctx context.Context, policy []byte, data []byte) error {
 	return p.updateVMs(func(vm *VM, opts vmOpts) error {
 		opts.policy = policy
 		opts.data = data
-		return vm.SetPolicyData(opts)
+		return vm.SetPolicyData(ctx, opts)
 	})
 }