--- a/internal/wasm/sdk/opa/loader/http/loader_test.go
+++ b/internal/wasm/sdk/opa/loader/http/loader_test.go
@@ -93,7 +93,7 @@ type testPolicyData struct {
 	updated chan struct{}
 }
 
-func (pd *testPolicyData) SetPolicyData(policy []byte, data *interface{}) error {
+func (pd *testPolicyData) SetPolicyData(_ context.Context, policy []byte, data *interface{}) error {
 	pd.Lock()
 	defer pd.Unlock()